+++ tests/testthat/test-GDALVector-class.R
@@ -39,3 +39,46 @@ test_that("class constructors work", {
 
     unlink(dsn)
 })
+
+test_that("setting ignored fields works", {
+    f <- system.file("extdata/ynp_fires_1984_2022.gpkg", package="gdalraster")
+    dsn <- file.path(tempdir(), basename(f))
+    file.copy(f, dsn, overwrite = TRUE)
+
+    lyr <- new(GDALVector, dsn)
+    expect_true(lyr$testCapability()$IgnoreFields)
+    feat <- lyr$getNextFeature()
+    expect_length(feat, 10)
+    lyr$setIgnoredFields("event_id")
+    feat <- lyr$getNextFeature()
+    expect_length(feat, 9)
+    expect_true(is.null(feat$event_id))
+    expect_false(is.null(feat$incid_name))
+    lyr$setIgnoredFields("")
+    feat <- lyr$getNextFeature()
+    expect_length(feat, 10)
+    lyr$setIgnoredFields(c("event_id", "map_id", "ig_year"))
+    feat <- lyr$getNextFeature()
+    expect_length(feat, 7)
+    lyr$setIgnoredFields("")
+    feat <- lyr$getNextFeature()
+    expect_length(feat, 10)
+
+    # geometry
+    lyr$returnGeomAs <- "WKT"
+    feat <- lyr$getNextFeature()
+    expect_length(feat, 11)
+    lyr$setIgnoredFields("OGR_GEOMETRY")
+    feat <- lyr$getNextFeature()
+    expect_length(feat, 10)
+    expect_true(is.null(feat$geom))
+    lyr$setIgnoredFields(c("event_id", "OGR_GEOMETRY"))
+    feat <- lyr$getNextFeature()
+    expect_length(feat, 9)
+    lyr$setIgnoredFields("")
+    feat <- lyr$getNextFeature()
+    expect_length(feat, 11)
+
+    lyr$close()
+    unlink(dsn)
+})